Move affordance.js to ts.
